% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MInt.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap}
\usage{
bootstrap(mfit, nboot = 10, seed = 1)
}
\arguments{
\item{mfit}{A MInt model object.}

\item{nboot}{The number of bootstraps to perform.}

\item{seed}{Random number generator seed.}
}
\value{
A MInt object.
}
\description{
This function bootstraps a model learned by \code{estimate} to
obtain confidence intervals on each parameter.
}
