% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MInt.R
\name{mint}
\alias{mint}
\title{Construct MInt object}
\usage{
mint(y, x, fmla = ~1)
}
\arguments{
\item{y}{A file path to the response matrix.}

\item{x}{A file path to the design matrix.}

\item{fmla}{An object of class ``\code{\link{formula}}'' (or one that can be
coerced to that class): a symbolic description of the model to be fitted.}
}
\value{
mint A MInt object.
}
\description{
Constructs a MInt object that maintains the data and parameter estimates for
the underlying Poisson-multivariate normal hierarchical model.
}
\examples{
x <- system.file("extdata", "x.txt", package="MInt");
y <- system.file("extdata", "y.txt", package="MInt");
m <- mint(y,x,fmla = ~feature1 + feature2)

}
