% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai.R
\name{topic_words_to_ai_prompt}
\alias{topic_words_to_ai_prompt}
\title{Generate prompt for topic naming}
\usage{
topic_words_to_ai_prompt(k, topic_names, n_words)
}
\arguments{
\item{k}{The number of topics}

\item{topic_names}{Pseudo-names used for the topics}

\item{n_words}{Number of words included in pseudo-names}
}
\value{
A character prompt
}
\description{
Construct a prompt to be used to name topics resulting from an topic model.
}
