% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap}
\usage{
bootstrap(mfit, nboot = 10, seed = 1)
}
\arguments{
\item{mfit}{A MInt model object.}

\item{nboot}{The number of bootstraps to perform.}

\item{seed}{Random number generator seed.}
}
\value{
A MInt object.

  We should export this at some point.
}
\description{
This function bootstraps a model learned by \code{estimate} to
obtain confidence intervals on each parameter.
}

