% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinT.R
\docType{package}
\name{MinT}
\alias{MinT}
\title{Find test statistic value from a given data and corresponding critical value}
\description{
This package finds test statistic value and
critical value of Min-T test for testing equality of means aginst
ordered alternatives in one-way ANOVA model with heteroscedasticity.
}
\author{
Anjana Mondal \email{mondal.anjana25@gmail.com}
}
